Add SDH_ResetCard function for resetting card.
Forcefully reset the SD card state to idle to prevent DAT3 voltage in an unexpected level due to transmission terminated. In DAT3 SD card detection mode, an unexpected IO state occurs. When the card is inserted and a low voltage level on DAT3 pin is detected. At this point, we issue a CMD0 command to the SD card to transition it into the IDLE mode, causing the DAT3 signal to return to a high voltage level.

Change-Id: Ib95dd777dfd9df548fc53a200bc70e5c81a7fe37